pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "dotenv>=0.9.9",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
]
